+++ crates/sncast/src/main.rs
@@ -12,6 +12,7 @@ use camino::Utf8PathBuf;
 use clap::{Parser, Subcommand};
 use sncast::helpers::configuration::CastConfig;
 use sncast::helpers::constants::{DEFAULT_ACCOUNTS_FILE, DEFAULT_MULTICALL_CONTENTS};
+use sncast::helpers::fee::PayableTransaction;
 use sncast::helpers::scarb_utils::{
     assert_manifest_path_exists, build, build_and_load_artifacts, get_package_metadata,
     get_scarb_metadata_with_deps, BuildConfig,
@@ -186,6 +187,8 @@ async fn run_async_command(
         Commands::Declare(declare) => {
             let provider = get_provider_and_verify_rpc_version(&config.url).await?;
 
+            declare.validate()?;
+
             let account = get_account(
                 &config.account,
                 &config.accounts_file,
@@ -204,16 +207,10 @@ async fn run_async_command(
                 },
             )
             .expect("Failed to build contract");
-            let mut result = starknet_commands::declare::declare(
-                &declare.contract,
-                declare.fee.max_fee,
-                &account,
-                declare.nonce,
-                &artifacts,
-                wait_config,
-            )
-            .await
-            .map_err(handle_starknet_command_error);
+            let mut result =
+                starknet_commands::declare::declare(declare, &account, &artifacts, wait_config)
+                    .await
+                    .map_err(handle_starknet_command_error);
 
             print_command_result("declare", &mut result, numbers_format, &output_format)?;
             Ok(())
@@ -262,6 +259,8 @@ async fn run_async_command(
         Commands::Invoke(invoke) => {
             let provider = get_provider_and_verify_rpc_version(&config.url).await?;
 
+            invoke.validate()?;
+
             let account = get_account(
                 &config.account,
                 &config.accounts_file,
@@ -270,13 +269,10 @@ async fn run_async_command(
             )
             .await?;
             let mut result = starknet_commands::invoke::invoke(
-                invoke.contract_address,
+                invoke.clone(),
                 get_selector_from_name(&invoke.function)
                     .context("Failed to convert entry point selector to FieldElement")?,
-                invoke.calldata,
-                invoke.fee.max_fee,
                 &account,
-                invoke.nonce,
                 wait_config,
             )
             .await
@@ -308,6 +304,8 @@ async fn run_async_command(
                 starknet_commands::multicall::Commands::Run(run) => {
                     let provider = get_provider_and_verify_rpc_version(&config.url).await?;
 
+                    run.validate()?;
+
                     let account = get_account(
                         &config.account,
                         &config.accounts_file,
@@ -315,13 +313,9 @@ async fn run_async_command(
                         config.keystore,
                     )
                     .await?;
-                    let mut result = starknet_commands::multicall::run::run(
-                        &run.path,
-                        &account,
-                        run.max_fee,
-                        wait_config,
-                    )
-                    .await;
+                    let mut result =
+                        starknet_commands::multicall::run::run(run.clone(), &account, wait_config)
+                            .await;
 
                     print_command_result(
                         "multicall run",
@@ -591,6 +585,8 @@ fn update_cast_config(config: &mut CastConfig, cli: &Cli) {
         ),
         clone_or_else!(cli.wait_timeout, config.wait_params.get_timeout()),
     );
+
+    config.url = clone_or_else!(cli.rpc_url, config.url);
 
     let url = match cli.command {
         Commands::Declare(declare::Declare { ref rpc_url, .. }) => rpc_url,