+++ crates/sncast/src/starknet_commands/declare.rs
@@ -1,13 +1,14 @@
 use anyhow::{anyhow, Context, Result};
-use clap::Args;
+use clap::{Args, ValueEnum};
 use scarb_api::StarknetContractArtifacts;
 use sncast::response::structs::DeclareResponse;
 use sncast::response::structs::Felt;
-use sncast::{apply_optional, handle_wait_for_tx, ErrorData, WaitForTx};
+use sncast::{apply_optional, handle_wait_for_tx, impl_payable_transaction, ErrorData, WaitForTx};
 use starknet::accounts::AccountError::Provider;
-use starknet::accounts::{ConnectedAccount, DeclarationV2};
+use starknet::accounts::{ConnectedAccount, DeclarationV2, DeclarationV3};
 
-use sncast::helpers::fee::FeeArgs;
+use sncast::helpers::error::token_not_supported_for_declaration;
+use sncast::helpers::fee::{FeeArgs, FeeSettings, FeeToken, PayableTransaction};
 use sncast::response::errors::StarknetCommandError;
 use starknet::core::types::FieldElement;
 use starknet::{
@@ -27,7 +28,7 @@ pub struct Declare {
     pub contract: String,
 
     #[clap(flatten)]
-    pub fee: FeeArgs,
+    pub fee_args: FeeArgs,
 
     /// Nonce of the transaction. If not provided, nonce will be set automatically
     #[clap(short, long)]
@@ -40,23 +41,42 @@ pub struct Declare {
     /// RPC provider url address; overrides url from snfoundry.toml
     #[clap(short = 'u', long = "url")]
     pub rpc_url: Option<String>,
+
+    /// Version of the declaration (can be inferred from fee token)
+    #[clap(short, long)]
+    pub version: Option<DeclareVersion>,
 }
+
+#[derive(ValueEnum, Debug, Clone)]
+pub enum DeclareVersion {
+    V2,
+    V3,
+}
+
+impl_payable_transaction!(Declare, token_not_supported_for_declaration,
+    DeclareVersion::V2 => FeeToken::Eth,
+    DeclareVersion::V3 => FeeToken::Strk
+);
 
 #[allow(clippy::too_many_lines)]
 pub async fn declare(
-    contract_name: &str,
-    max_fee: Option<FieldElement>,
+    declare: Declare,
     account: &SingleOwnerAccount<&JsonRpcClient<HttpTransport>, LocalWallet>,
-    nonce: Option<FieldElement>,
     artifacts: &HashMap<String, StarknetContractArtifacts>,
     wait_config: WaitForTx,
 ) -> Result<DeclareResponse, StarknetCommandError> {
-    let contract_name: String = contract_name.to_string();
+    let fee_settings = declare
+        .fee_args
+        .clone()
+        .fee_token(declare.token_from_version())
+        .try_into_fee_settings(account.provider(), account.block_id())
+        .await?;
+
     let contract_artifacts =
         artifacts
-            .get(&contract_name)
+            .get(&declare.contract)
             .ok_or(StarknetCommandError::ContractArtifactsNotFound(
-                ErrorData::new(contract_name),
+                ErrorData::new(declare.contract),
             ))?;
 
     let contract_definition: SierraClass = serde_json::from_str(&contract_artifacts.sierra)
@@ -68,14 +88,36 @@ pub async fn declare(
         .class_hash()
         .map_err(anyhow::Error::from)?;
 
-    let declaration = account.declare_v2(
-        Arc::new(contract_definition.flatten().map_err(anyhow::Error::from)?),
-        casm_class_hash,
-    );
+    let declared = match fee_settings {
+        FeeSettings::Eth { max_fee } => {
+            let declaration = account.declare_v2(
+                Arc::new(contract_definition.flatten().map_err(anyhow::Error::from)?),
+                casm_class_hash,
+            );
 
-    let declaration = apply_optional(declaration, max_fee, DeclarationV2::max_fee);
-    let declaration = apply_optional(declaration, nonce, DeclarationV2::nonce);
-    let declared = declaration.send().await;
+            let declaration = apply_optional(declaration, max_fee, DeclarationV2::max_fee);
+            let declaration = apply_optional(declaration, declare.nonce, DeclarationV2::nonce);
+
+            declaration.send().await
+        }
+        FeeSettings::Strk {
+            max_gas,
+            max_gas_unit_price,
+        } => {
+            let declaration = account.declare_v3(
+                Arc::new(contract_definition.flatten().map_err(anyhow::Error::from)?),
+                casm_class_hash,
+            );
+
+            let declaration = apply_optional(declaration, max_gas, DeclarationV3::gas);
+            let declaration =
+                apply_optional(declaration, max_gas_unit_price, DeclarationV3::gas_price);
+            let declaration = apply_optional(declaration, declare.nonce, DeclarationV3::nonce);
+
+            declaration.send().await
+        }
+    };
+
     match declared {
         Ok(result) => handle_wait_for_tx(
             account.provider(),